Extract Fieldset and ScmToolRadioGroup
