+++ src/components/forms/index.js
@@ -3,3 +3,4 @@ export { default as TextField } from './TextField';
 export { default as Radio } from './Radio';
 export { default as Checkbox } from './Checkbox';
 export { default as HelpText } from './HelpText';
+export { default as Fieldset } from './Fieldset';
