+++ src/components/index.js
@@ -19,5 +19,5 @@ export { default as TwoColumnLayout } from './TwoColumnLayout';
 export { default as ResponsiveSpacer } from './ResponsiveSpacer';
 export { default as TextLink } from './TextLink';
 export { default as ContentHeader } from './ContentHeader';
-export { Radio, TextField, Button, Checkbox, HelpText } from './forms';
+export { Radio, TextField, Button, Checkbox, HelpText, Fieldset } from './forms';
 export { ButtonLinkCallToAction } from './CallToAction';
