+++ src/contactFormConstants.js
@@ -9,34 +9,3 @@ export const FORM_NAMES = {
   getInstanceExtended: 'submit-extended-get-instance-form',
   requestDemo: 'submit-extended-get-demo-form',
 };
-
-export const SCM_TOOLS = [{
-  value: 'github-cloud',
-  label: 'GitHub Cloud',
-  supported: true,
-}, {
-  value: 'github-on-prem',
-  label: 'GitHub On-prem',
-  supported: false,
-}, {
-  value: 'gitlab-cloud',
-  label: 'Gitlab Cloud',
-  supported: false,
-}, {
-  value: 'gitlab-on-prem',
-  label: 'Gitlab On-prem',
-  supported: false,
-}, {
-  value: 'bitbucket-cloud',
-  label: 'Bitbucket Cloud',
-  supported: false,
-}, {
-  value: 'bitbucket-server',
-  label: 'Bitbucket Server',
-  supported: false,
-}, {
-  value: 'other',
-  label: 'Other',
-  supported: false,
-}];
-
